[PM-28540] Add AuthenticatorSharedMocks target
