+++ AuthenticatorShared/Core/Platform/Services/API/Fixtures/APITestData+Authenticator.swift
@@ -1,12 +1,12 @@
 import TestHelpers
 
-/// `APITestData` helpers that load the resource from the `AuthenticatorShared` bundle.
+/// `APITestData` helpers that load the resource from the `AuthenticatorSharedMocks` bundle.
 extension APITestData {
     static func loadFromBundle(resource: String, extension: String) -> APITestData {
-        loadFromBundle(resource: resource, extension: `extension`, bundle: .authenticatorShared)
+        loadFromBundle(resource: resource, extension: `extension`, bundle: .authenticatorSharedMocks)
     }
 
     static func loadFromJsonBundle(resource: String) -> APITestData {
-        loadFromJsonBundle(resource: resource, bundle: .authenticatorShared)
+        loadFromJsonBundle(resource: resource, bundle: .authenticatorSharedMocks)
     }
 }
