+++ AuthenticatorShared/Core/Platform/Services/API/Fixtures/AuthenticatorSharedMocksBundleFinder.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+class AuthenticatorSharedMocksBundleFinder {
+    static let bundle = Bundle(for: AuthenticatorSharedMocksBundleFinder.self)
+}
+
+public extension Bundle {
+    /// The bundle for the `AuthenticatorSharedMocks` target.
+    static let authenticatorSharedMocks = AuthenticatorSharedMocksBundleFinder.bundle
+}
