+++ AuthenticatorShared/Core/Vault/Services/Fixtures/ImportTestData.swift
@@ -1,12 +1,14 @@
 import Foundation
 
+// swiftlint:disable missing_docs
+
 /// A type that wraps import format fixture data for use in tests.
 ///
-struct ImportTestData {
-    let data: Data
+public struct ImportTestData {
+    public let data: Data
 
     static func loadFromBundle(resource: String, extension: String) -> ImportTestData {
-        let bundle = Bundle(for: BitwardenTestCase.self)
+        let bundle = AuthenticatorSharedMocksBundleFinder.bundle
         guard let url = bundle.url(forResource: resource, withExtension: `extension`) else {
             fatalError("Unable to locate file \(resource).\(`extension`) in the bundle.")
         }
@@ -22,9 +24,11 @@ struct ImportTestData {
     }
 }
 
-extension ImportTestData {
+public extension ImportTestData {
     static let bitwardenJson = loadFromJsonBundle(resource: "BitwardenExport")
     static let lastpassJson = loadFromJsonBundle(resource: "LastpassExport")
     static let raivoJson = loadFromJsonBundle(resource: "RaivoExport")
     static let twoFasJson = loadFromBundle(resource: "TwoFasExport", extension: "2fas")
 }
+
+// swiftlint:enable missing_docs
