+++ AuthenticatorShared/Core/Vault/Services/Importers/BitwardenImporterTests.swift
@@ -1,3 +1,4 @@
+import AuthenticatorSharedMocks
 import XCTest
 
 @testable import AuthenticatorShared
