+++ AuthenticatorShared/Core/Vault/Services/TOTP/TOTPExpirationCalculatorTests.swift
@@ -1,3 +1,4 @@
+import AuthenticatorSharedMocks
 import BitwardenKitMocks
 import XCTest
